+++ fiddles/angular2/fiddle-0041-JqueryCSV/app/meta.ts
@@ -1,0 +1,7 @@
+var metadata = {
+     fiddleHeader: 'Angular 2 - JQuery CSV',
+     urls: {
+          github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0041-JQueryCSV'
+     },
+     consoleTag: 'H O U S E ~ f i d d l e s'
+};