+++ fiddles/aurelia/fiddle-0004-HTTPServices/app.js
@@ -1,3 +1,15 @@
+import { YahooChartApiService } from './yahoo-chart-api.service';
+import { inject } from "aurelia-framework";
+
+@inject(YahooChartApiService)
 export class App {
-  message = 'hello world';
+   constructor(yahooChartApiService) {
+       this.service = yahooChartApiService;
+   }
+
+    activate(){
+        this.service.getAll().then(data => {
+            this.json =	data;
+        });
+    }
 }