+++ fiddles/svg/fiddle-0000-Template/app.js
@@ -1,0 +1,3 @@
+$(document).ready(function () {
+    $('#canvas').height(window.innerHeight);
+});